selenium_m/requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
Faker==4.1.2
loguru==0.6.0
pandas==1.0.3
py==1.9.0
pytest==7.2.0
pytest-html==3.2.0
requests==2.24.0
selenium==4.8.2
xlrd==1.2.0
xlwt==1.3.0